+++ src/Driver/Mysql.php
@@ -31,6 +31,7 @@ class Mysql extends Driver
     /**
      * Gets the executable command to export the database
      * @return string
+     * @uses getBinary()
      * @uses getConfig()
      * @uses $auth
      */
@@ -47,6 +48,7 @@ class Mysql extends Driver
     /**
      * Gets the executable command to import the database
      * @return string
+     * @uses getBinary()
      * @uses getConfig()
      * @uses $auth
      */