+++ src/Driver/Postgres.php
@@ -47,6 +47,7 @@ class Postgres extends Driver
     /**
      * Gets the executable command to export the database
      * @return string
+     * @uses getBinary()
      * @uses getDbnameAsString()
      */
     protected function _exportExecutable(): string
@@ -57,6 +58,7 @@ class Postgres extends Driver
     /**
      * Gets the executable command to import the database
      * @return string
+     * @uses getBinary()
      * @uses getDbnameAsString()
      */
     protected function _importExecutable(): string