+++ src/Driver/Sqlite.php
@@ -24,6 +24,7 @@ class Sqlite extends Driver
     /**
      * Gets the executable command to export the database
      * @return string
+     * @uses getBinary()
      * @uses getConfig()
      */
     protected function _exportExecutable(): string
@@ -34,6 +35,7 @@ class Sqlite extends Driver
     /**
      * Gets the executable command to import the database
      * @return string
+     * @uses getBinary()
      * @uses getConfig()
      */
     protected function _importExecutable(): string